Fixes progress bar indicator

+++ src/source.cpp
@@ -51,9 +51,9 @@ public:
 void PrintProgress(float progress) {
   printf("[");
   for (int i = 0; i < 40; ++ i) {
-    if (i <  40.0f*progress) printf("=");
-    if (i == 40.0f*progress) printf(">");
-    if (i >  40.0f*progress) printf(" ");
+    if (i <  static_cast<int>(40.0f*progress)) printf("=");
+    if (i == static_cast<int>(40.0f*progress)) printf(">");
+    if (i >  static_cast<int>(40.0f*progress)) printf(" ");
   }
   // leading spaces in case of terminal/text corruption
   printf("] %0.1f%%   \r", progress*100.0f);
